docs: Remove mention of unsecure ws connection
Removed the mention of unsecure WebSockets connection and only left the secure one. Reason for this is that you can connect to a secure websocket from both http and https, but you can only connect to a unsecure one from http and not https, unless you want the user to accept "unsecure connection" warnings.