fix: dont transform listening addrs (#195)
This fixes an unncessary change from #194. The transport
should not be modifying addresses. This was being done to overcome
an issue with mafmt not supporting all valid webrtc-star addresses,
however that was fixed and this hack wasn't reverted. This
reverts the hack to keep consistency with transport listening and
getAddrs behavior.